backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
sqlalchemy[asyncio]>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.0
httpx>=0.27.0
google-genai>=1.0
python-dotenv>=1.0
PyJWT[crypto]>=2.8.0
